+++ htmlresources/symbolstreeview/js/startup.js
@@ -1,0 +1,5 @@
+var symbolsTreeView;
+
+$(function() {
+    symbolsTreeView = new SymbolsTreeView();
+});